Bug fixes for damage powerup

+++ EECS494-P3/Assets/Scripts/Enemy/BossHasHealth.cs
@@ -36,7 +36,7 @@ public class BossHasHealth : HasHealth
     }
 
 
-    public override void AlterHealth(int healthDelta)
+    public override void AlterHealth(float healthDelta)
     {
         base.AlterHealth(healthDelta);
         StartCoroutine(FlashRed());
