+++ EECS494-P3/Assets/Scripts/Enemy/HasEnemyHealth.cs
@@ -19,7 +19,7 @@ public class HasEnemyHealth : HasHealth {
         }
     }
 
-    public override void AlterHealth(int healthDelta) {
+    public override void AlterHealth(float healthDelta) {
         base.AlterHealth(healthDelta);
         StartCoroutine(FlashRed());
         if (health <= 0) {
