+++ EECS494-P3/Assets/Scripts/HasHealth.cs
@@ -12,7 +12,7 @@ public class HasHealth : MonoBehaviour {
     }
 
     // Update is called once per frame
-    public virtual void AlterHealth(int healthDelta) {
+    public virtual void AlterHealth(float healthDelta) {
         health += healthDelta;
     }
